+++ src/network/CMakeLists.txt
@@ -8,14 +8,6 @@ target_link_libraries(websocket_utils
 
 add_library(mission_control_interface STATIC
   MissionControlProtocol.cpp)
+target_link_libraries(mission_control_interface utils)
 set_target_properties(mission_control_interface PROPERTIES POSITION_INDEPENDENT_CODE ON)
-if(WITH_CV)
-  target_sources(mission_control_interface PRIVATE
-	${CMAKE_CURRENT_LIST_DIR}/MissionControlStreaming.cpp
-	../base64/base64_img.cpp)
-  target_link_libraries(mission_control_interface ${OpenCV_LIBRARIES})
-else()
-  target_sources(mission_control_interface PRIVATE
-	${CMAKE_CURRENT_LIST_DIR}/MissionControlStreamingStub.cpp)
-endif()
